requirements.txt
altair==5.4.1
appnope==0.1.4
asn1crypto==1.5.1
asttokens==2.4.1
attrs==24.2.0
blinker==1.8.2
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
cryptography==43.0.3
debugpy==1.8.7
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
gitdb==4.0.11
GitPython==3.1.43
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.10
importlib_metadata==7.2.1
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
narwhals==1.10.0
nest_asyncio==1.6.0
numpy==1.26.4
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==24.2
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
polars==1.11.0
prompt_toolkit==3.0.48
protobuf==5.27.5
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pycparser==2.22
pydeck==0.8.0b4
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
PySocks==1.7.1
python-dateutil==2.9.0
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rich==13.9.3
rpds-py==0.20.0
ruff==0.7.1
setuptools==75.1.0
six==1.16.0
smmap==5.0.0
snowflake-connector-python==3.12.2
sortedcontainers==2.4.0
stack-data==0.6.2
streamlit==1.39.0
tenacity==9.0.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
validators==0.34.0
watchdog==5.0.3
wcwidth==0.2.13
wheel==0.44.0
widgetsnbextension==4.0.13
zipp==3.20.2
zstandard==0.23.0
